+++ Reflection/src/utopia/reflection/component/reach/input/EditableTextLabel.scala
@@ -1,27 +1,27 @@
 package utopia.reflection.component.reach.input
 
+import java.awt.Toolkit
+import java.awt.datatransfer.{Clipboard, ClipboardOwner, DataFlavor, StringSelection, Transferable}
+import java.awt.event.KeyEvent
+
 import utopia.flow.datastructure.mutable.PointerWithEvents
-import utopia.flow.event.Changing
 import utopia.flow.util.TimeExtensions._
+import utopia.flow.util.StringExtensions._
 import utopia.genesis.color.Color
-import utopia.genesis.handling.Actor
-import utopia.genesis.shape.shape2D.{Bounds, Line, Point}
-import utopia.genesis.util.Drawer
+import utopia.genesis.event.{KeyStateEvent, KeyTypedEvent}
+import utopia.genesis.handling.{Actor, KeyStateListener, KeyTypedListener}
 import utopia.inception.handling.HandlerType
-import utopia.inception.handling.immutable.Handleable
 import utopia.reflection.component.drawing.immutable.TextDrawContext
-import utopia.reflection.component.drawing.template.CustomDrawer
-import utopia.reflection.component.drawing.template.DrawLevel.Foreground
 import utopia.reflection.component.reach.hierarchy.ComponentHierarchy
-import utopia.reflection.component.reach.label.ViewTextLabel
-import utopia.reflection.component.reach.template.{Focusable, MutableCustomDrawReachComponent, MutableFocusable, ReachComponentWrapper}
+import utopia.reflection.component.reach.template.{MutableCustomDrawReachComponent, MutableFocusable}
 import utopia.reflection.component.template.text.MutableTextComponent
-import utopia.reflection.event.{FocusChangeEvent, FocusChangeListener, FocusListener, FocusStateTracker}
+import utopia.reflection.event.{FocusChangeEvent, FocusChangeListener, FocusListener}
 import utopia.reflection.localization.LocalizedString
 import utopia.reflection.text.{FontMetricsContext, MeasuredText, Regex}
 import utopia.reflection.localization.LocalString._
 
 import scala.concurrent.duration.{Duration, FiniteDuration}
+import scala.util.{Failure, Success, Try}
 
 /**
   * Used for requesting user input in text format
@@ -47,8 +47,7 @@ class EditableTextLabel(override val parentHierarchy: ComponentHierarchy,
 	}
 	private val caretIndexPointer = new PointerWithEvents(textPointer.value.length - 1)
 	private val caretVisibilityPointer = new PointerWithEvents(false)
-	
-	private var selectionStartIndex: Option[Int] = None
+	private val selectedRangePointer = new PointerWithEvents[Option[Range]](None)
 	
 	
 	// INITIAL CODE	-------------------------------
@@ -61,6 +60,10 @@ class EditableTextLabel(override val parentHierarchy: ComponentHierarchy,
 	def hasFocus = FocusHandler.hasFocus
 	
 	def caretIndex = caretIndexPointer.value
+	
+	def selectedRange = selectedRangePointer.value
+	
+	def selectedText = selectedRange.map { textPointer.value(_) }.filterNot { _.isEmpty }
 	
 	def text_=(newText: String) = textPointer.value = newText
 	
@@ -86,20 +89,13 @@ class EditableTextLabel(override val parentHierarchy: ComponentHierarchy,
 	
 	// OTHER	----------------------------------
 	
-	/*
-	private def caretsFrom(text: String) =
+	def moveCaretToEnd() = caretIndexPointer.value = textPointer.value.length - 1
+	
+	private def insertToCaret(text: String) =
 	{
-		val lineHeight = wrapped.singleLineTextHeight
-		val lines = text.linesIterator.toVector
-		lines.zipWithIndex.flatMap { case (line, index) =>
-			val topY = (index * lineHeight) + ((index - 1) max 0) * wrapped.betweenLinesMargin
-			val bottomY = topY + lineHeight
-			line.indices.map { i =>
-				val x = wrapped.textWidthWith(line.take(i + 1))
-				Line(Point(x, topY), Point(x, bottomY))
-			}
-		}
-	}*/
+		val (before, after) = textPointer.value.splitAt(caretIndex)
+		textPointer.value = before + text + after
+	}
 	
 	
 	// NESTED	----------------------------------
@@ -117,7 +113,10 @@ class EditableTextLabel(override val parentHierarchy: ComponentHierarchy,
 		{
 			hasFocus = event.hasFocus
 			if (hasFocus)
+			{
+				moveCaretToEnd()
 				CaretBlinker.resetCounter()
+			}
 		}
 	}
 	
@@ -146,5 +145,77 @@ class EditableTextLabel(override val parentHierarchy: ComponentHierarchy,
 		// OTHER	------------------------------
 		
 		def resetCounter() = passedDuration = Duration.Zero
+	}
+	
+	private object KeyListener extends KeyTypedListener with KeyStateListener with ClipboardOwner
+	{
+		// ATTRIBUTES	--------------------------
+		
+		override val keyStateEventFilter = KeyStateEvent.wasPressedFilter
+		
+		
+		// COMPUTED	------------------------------
+		
+		// NB: May throw
+		private def clipBoard = Toolkit.getDefaultToolkit.getSystemClipboard
+		
+		
+		// IMPLEMENTED	--------------------------
+		
+		override def onKeyTyped(event: KeyTypedEvent) =
+		{
+			// Inserts the typed character into the string (if accepted by the content filter)
+			if (inputFilter.forall { _(event.typedChar.toString) })
+				insertToCaret(event.typedChar.toString)
+		}
+		
+		// TODO: Handle arrow key press events
+		override def onKeyState(event: KeyStateEvent) =
+		{
+			// Inserts a line-break on enter (if enabled)
+			if (allowLineBreaks && event.index == KeyEvent.VK_ENTER)
+				insertToCaret("\n")
+			// Listens to shortcut keys (ctrl + C, V or X)
+			else if (event.keyStatus.control)
+			{
+				if (event.index == KeyEvent.VK_C)
+					selectedText.filterNot { _.isEmpty }.foreach { selected => Try { clipBoard.setContents(
+						new StringSelection(selected), this) }
+					}
+				else if (event.index == KeyEvent.VK_V)
+					Try {
+						Option(clipBoard.getContents(null)).foreach { pasteContent =>
+							if (pasteContent.isDataFlavorSupported(DataFlavor.stringFlavor))
+							{
+								val rawPasteText = pasteContent.getTransferData(DataFlavor.stringFlavor).toString
+								inputFilter match
+								{
+									case Some(filter) => insertToCaret(filter.filter(rawPasteText))
+									case None => insertToCaret(rawPasteText)
+								}
+							}
+						}
+					}
+				else if (event.index == KeyEvent.VK_X)
+					selectedRange.foreach { range =>
+						val textBefore = textPointer.value
+						val (cut, remain) = textBefore.cut(range)
+						if (cut.nonEmpty)
+						{
+							// Copies the cut content to the clip board. Will not remove the text if copy failed.
+							Try { clipBoard.setContents(new StringSelection(cut), this) } match
+							{
+								case Success(_) => textPointer.value = remain
+								case Failure(_) => ()
+							}
+						}
+					}
+			}
+		}
+		
+		// Only listens to key events while focused
+		override def allowsHandlingFrom(handlerType: HandlerType) = hasFocus
+		
+		override def lostOwnership(clipboard: Clipboard, contents: Transferable) = ()
 	}
 }